Use bigint for version number so date stamps can be used

+++ zf1/Akrabat/Db/Schema/Manager.php
@@ -59,7 +59,7 @@ class Akrabat_Db_Schema_Manager
         } catch (Zend_Db_Exception $e) {
             // exception means that the schema version table doesn't exist, so create it
             $createSql = "CREATE TABLE $schemaVersionTableName ( 
-                version int NOT NULL,
+                version bigint NOT NULL,
                 PRIMARY KEY (version)
             )";
             $this->_db->query($createSql);
